Updated project to Swift 5.0 using migration tools.
I found this project in some Google exploration, and I found that
it was crashing on startup in my current version of macOS. After
cloning and opening in Xcode, it offered to migrate the code to a
more modern version of Swift.

Given this was originally written in Swift 3.0, I had to download
an older version, Xcode 10.1, to first migrate the app to Swift
4.0, then open in the current release of Xcode, 10.2.1, to
migrate to Swift 5.0.

The project builds just fine, and I can confirm it runs on my
system in macOS 10.14.5 Mojave.

Resolves: #3